requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pandas
sentence-transformers
faiss-cpu
scikit-learn
numpy
python-multipart